CFG Assigment 4 - Library API/requirements.txt
Flask
MySQL-python
requests